modified code as per pr suggestions
