+++ src/common/components/chat/message-formatting.ts
@@ -1,0 +1,7 @@
+export const processMessageForDisplay = (message: string): string => message
+    // quick sanitization
+    .replaceAll(/</g, '&lt;')
+    .replaceAll(/>/g, '&gt;')
+    // bold texts wrapped **like this**
+    .replaceAll(/\*\*(.*?)\*\*/g, '<b>$1</b>')
+    .replaceAll(/\n/g, '<br>');
